RAG/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn>=0.23.0,<0.31.0
python-dotenv>=1.0.1,<2.0.0
openai>=1.30.0,<2.0.0
numpy>=1.26.4,<3.0.0
pydantic>=2.6.0,<3.0.0
pypdf>=4.2.0,<5.0.0
beautifulsoup4>=4.12.3,<5.0.0

